requirements.txt
azure-storage-blob==12.13.1
flask==2.2.2
from-root==1.0.2
pip-chill==1.0.1
pyyaml==6.0
scikit-learn==1.0.2
wincertstore==0.2
